+++ tester/utils/test-patterns/agent-test.hh
@@ -69,7 +69,8 @@ public:
 	void operator()() final {
 		configureAgent();
 		mRegistrarDb = std::make_shared<RegistrarDb>(mRoot, mConfigManager);
-		mAgent = std::make_shared<Agent>(mRoot, mConfigManager, mAuthDb, mRegistrarDb);
+		mAgent = std::make_shared<Agent>(mRoot, mConfigManager, mAuthDb, mRegistrarDb,
+		                                 SpacesStore::make(mRoot, mConfigManager, nullptr));
 		onAgentConfigured();
 		if (mRunAgent) {
 			mAgent->start("", "");